+++ RaftNET.Tests/ReplicationTests/Stop.cs
@@ -1,3 +1,3 @@
 ﻿namespace RaftNET.Tests.ReplicationTests;
 
-public record Stop(int Id);
+public record Stop(ulong Id);
