+++ RaftNET.Tests/ReplicationTests/WaitLog.cs
@@ -1,3 +1,3 @@
 ﻿namespace RaftNET.Tests.ReplicationTests;
 
-public record WaitLog(List<int> IntIds);
+public record WaitLog(List<ulong> IntIds);
